+++ Depo/Depo/FailurePurchasingView.m
@@ -132,7 +132,7 @@
     return self;
 }
 
-- (NSString *) quotaCalculator:(long) quota {
+- (NSString *) quotaCalculator:(float) quota {
     int quotInt = (int)( quota/(1024*1024*1024));
     NSString *quotStr = [NSString stringWithFormat:@"%d",quotInt];
     return quotStr;
